Move dialogs into their own folder
